+++ lib/lightning_web/components/layout_components.ex
@@ -21,19 +21,10 @@ defmodule LightningWeb.LayoutComponents do
         {__ENV__.module, __ENV__.function, __ENV__.file, __ENV__.line}
       ) %>
     <% else %>
-      <%= if assigns[:projects] do %>
-        <Menu.projects_dropdown
-          projects={assigns[:projects]}
-          selected_project={assigns[:project]}
-        />
-      <% else %>
-        <div class="p-2 mb-4 mt-4 text-center text-primary-300 bg-primary-800">
-          <span class="inline-block align-middle text-sm">
-            You don't have access to any projects
-          </span>
-        </div>
-      <% end %>
-
+      <Menu.projects_dropdown
+        projects={assigns[:projects] || []}
+        selected_project={assigns[:project]}
+      />
       <%= if assigns[:project] do %>
         <Menu.project_items
           project_id={@project.id}
@@ -64,8 +55,10 @@ defmodule LightningWeb.LayoutComponents do
     """
   end
 
-  attr :current_user, Lightning.Accounts.User
+  attr :current_user, Lightning.Accounts.User, default: nil
   attr :socket, Phoenix.LiveView.Socket
+  attr :breadcrumbs, :list, default: []
+  attr :project, :map, default: nil
   slot :title
   slot :period
   slot :description
@@ -89,13 +82,58 @@ defmodule LightningWeb.LayoutComponents do
     ~H"""
     <div class="flex-none bg-white shadow-sm">
       <div class={[@title_class, @title_height]}>
-        <h1 class="text-3xl font-bold text-secondary-900 flex items-center">
-          <%= if assigns[:title], do: render_slot(@title) %>
-        </h1>
-        <%= if assigns[:period] do %>
-          <span class="ml-2 mt-3 text-xs">
-            <%= render_slot(@period) %>
-          </span>
+        <%= if @current_user do %>
+          <nav class="flex" aria-label="Breadcrumb">
+            <ol role="list" class="flex items-center space-x-4">
+              <li>
+                <div>
+                  <a href="#" class="text-gray-400 hover:text-gray-500">
+                    <svg
+                      class="h-5 w-5 flex-shrink-0"
+                      viewBox="0 0 20 20"
+                      fill="currentColor"
+                      aria-hidden="true"
+                    >
+                      <path
+                        fill-rule="evenodd"
+                        d="M9.293 2.293a1 1 0 011.414 0l7 7A1 1 0 0117 11h-1v6a1 1 0 01-1 1h-2a1 1 0 01-1-1v-3a1 1 0 00-1-1H9a1 1 0 00-1 1v3a1 1 0 01-1 1H5a1 1 0 01-1-1v-6H3a1 1 0 01-.707-1.707l7-7z"
+                        clip-rule="evenodd"
+                      />
+                    </svg>
+                    <span class="sr-only">Home</span>
+                  </a>
+                </div>
+              </li>
+
+              <%!-- If a project is scoped, we automatically generate the base crumbs --%>
+              <%= if @project do %>
+                <.breadcrumb path="/projects">
+                  <:label>Projects</:label>
+                </.breadcrumb>
+                <.breadcrumb path={"/projects/#{@project.id}/w"}>
+                  <:label><%= @project.name %></:label>
+                </.breadcrumb>
+              <% end %>
+
+              <%!-- If breamcrumbs are passed manually, we generate those --%>
+              <%= for {label, path} <- @breadcrumbs do %>
+                <.breadcrumb path={path}>
+                  <:label><%= label %></:label>
+                </.breadcrumb>
+              <% end %>
+
+              <%!-- And finally, we always show the page title --%>
+              <.breadcrumb>
+                <:label>
+                  <%= if assigns[:title], do: render_slot(@title) %>
+                </:label>
+              </.breadcrumb>
+            </ol>
+          </nav>
+        <% else %>
+          <h1 class="text-3xl font-bold text-secondary-900 flex items-center">
+            <%= if assigns[:title], do: render_slot(@title) %>
+          </h1>
         <% end %>
         <div class="grow"></div>
         <%= if assigns[:inner_block], do: render_slot(@inner_block) %>
@@ -137,11 +175,6 @@ defmodule LightningWeb.LayoutComponents do
           </.dropdown>
         <% end %>
       </div>
-      <%= if Enum.any?(assigns[:description]) do %>
-        <div class={[@title_class, "h-6 text-sm"]}>
-          <%= render_slot(@description) %>
-        </div>
-      <% end %>
     </div>
     """
   end
@@ -174,6 +207,43 @@ defmodule LightningWeb.LayoutComponents do
         </div>
       </div>
     </nav>
+    """
+  end
+
+  attr :path, :string, default: nil
+  slot :label
+
+  def breadcrumb(assigns) do
+    ~H"""
+    <li>
+      <div class="flex items-center">
+        <svg
+          class="h-5 w-5 flex-shrink-0 text-gray-400"
+          viewBox="0 0 20 20"
+          fill="currentColor"
+          aria-hidden="true"
+        >
+          <path
+            fill-rule="evenodd"
+            d="M7.21 14.77a.75.75 0 01.02-1.06L11.168 10 7.23 6.29a.75.75 0 111.04-1.08l4.5 4.25a.75.75 0 010 1.08l-4.5 4.25a.75.75 0 01-1.06-.02z"
+            clip-rule="evenodd"
+          />
+        </svg>
+        <%= if @path do %>
+          <.link
+            patch={@path}
+            class="flex ml-4 text-sm font-medium text-gray-500 hover:text-gray-700"
+            aria-current="page"
+          >
+            <%= if assigns[:label], do: render_slot(@label) %>
+          </.link>
+        <% else %>
+          <span class="flex ml-4 text-sm font-medium text-gray-500">
+            <%= if assigns[:label], do: render_slot(@label) %>
+          </span>
+        <% end %>
+      </div>
+    </li>
     """
   end
 end