+++ lib/lightning_web/components/layouts.ex
@@ -4,4 +4,7 @@ defmodule LightningWeb.Layouts do
   use LightningWeb, :html
 
   embed_templates "layouts/*"
+
+  slot :header_tags
+  def root(assigns)
 end