+++ lib/lightning_web/live/audit_live/index.ex
@@ -29,7 +29,7 @@ defmodule LightningWeb.AuditLive.Index do
     else
       {:ok,
        put_flash(socket, :nav, :no_access)
-       |> push_redirect(to: "/")}
+       |> push_redirect(to: "/projects")}
     end
   end