+++ test/lightning_web/live/profile_live_test.exs
@@ -87,7 +87,7 @@ defmodule LightningWeb.ProfileLiveTest do
         |> render_submit()
         |> follow_redirect(conn)
 
-      assert "/" = redirected_path = redirected_to(conn, 302)
+      assert "/projects" = redirected_path = redirected_to(conn, 302)
 
       html =
         get(recycle(conn), redirected_path)